// src/app/model/anthropometric-data.ts
export interface Anthropometrics {
    height: string;
    sittingHeight: string;
    shoulderHeight: string;
    lowerLegLength: string;
    hipBreadth: string;
    elbowHeight: string;
    eyeHeightStanding: string;
    elbowHeightStanding: string;
    thighClearance: string;
    eyeHeight: string;
    shoulderBreadth: string;
    kneeHeight: string;
    personEmail: string;
}